libdirect: os: linux: thread: Fix passing of null pointer to prctl
prctl(PR_GET_NAME, arg2, ...) expects a pointer to a 16 byte buffer
in arg2. Currently the code is checking if thread->name is null
and then calling prctl(PR_GET_NAME, thread->name, ...) if it is.
This results in a null pointer always being passed to prctl().

Put a 16 byte buffer on the stack, pass that to prctl() if thread->name
is null, then copy the result into thread->name.