system: drmkms: Probe for panel rotation
On some embedded systems the LCD panel is mounted in the chassis
upside down. To render the image correctly it needs to be rotated
in userspace.

The kernel part of DRM/KMS exposes the panel orientation via a
connector property. So grab the orientation for the panel from the
kernel and present that as the screen rotation so that the image
is flipped when blitted.